requirements.txt
pandas
pygwalker
pandasai
python-dotenv